requirements_dev.txt
azure-core==1.29.4
azure-storage-blob==12.18.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
packaging==23.1
pluggy==1.3.0
pycparser==2.21
pytest==7.4.2
pytest-mock==3.11.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.5
